Clean up hook module imports
